split js into modules
