+++ index.html
@@ -25,7 +25,10 @@
   <!-- App -->
   <link rel="stylesheet" href="/static/css/main.css"/>
   <script src="/static/js/localization.en-US.js" async></script>
-  <script src="/static/js/main.js" defer></script>
+  <script type="module" src="/static/js/ui.js"></script>
+  <script type="module" src="/static/js/plots.js"></script>
+  <script type="module" src="/static/js/data.js"></script>
+  <script type="module" src="/static/js/main.js"></script>
 </head>
 
 <body>
